+++ src/components/Code/index.js
@@ -1,23 +1,34 @@
 import React from 'react';
 
-const inlineCode = (children, testSection) => {
-  return (
-    <code
-      className="oui-code"
-      data-test-section={ testSection }>
-      { children }
-    </code>
-  );
-};
+import * as highlight from 'highlight.js';
 
-const blockCode = (children, testSection) => {
-  return (
-    <pre
-      className="oui-pre"
-      data-test-section={ testSection }>
-      <code>{ children }</code>
-    </pre>
-  );
+const highlightCode = (code, isHighlighted, className, testSection) => {
+  let codeElement = null;
+
+  if (isHighlighted) {
+    // Code that uses syntax highlighting needs to have
+    // `dangerouslySetInnerHTML` set so that the HTML returned is displayed.
+    require('highlight.js/styles/docco.css');
+    codeElement = (
+      /* eslint-disable react/no-danger */
+      <code
+        className={ className }
+        data-test-section={ testSection }
+        dangerouslySetInnerHTML={ { __html: highlight.highlightAuto(code).value } }>
+      </code>
+      /* eslint-enable react/no-danger */
+    );
+  } else {
+    codeElement = (
+      <code
+        className={ className }
+        data-test-section={ testSection }>
+        { code }
+      </code>
+    );
+  }
+
+  return codeElement;
 };
 
 /**
@@ -27,15 +38,23 @@ const blockCode = (children, testSection) => {
  */
 const Code = (props) => {
   if (props.type === 'inline') {
-    return inlineCode(props.children, props.testSection);
+    return highlightCode(props.children, props.isHighlighted, 'oui-code', props.testSection);
   }
 
-  return blockCode(props.children, props.testSection);
+  return (
+    <pre
+      className="oui-pre"
+      data-test-section={ props.testSection }>
+      { highlightCode(props.children, props.isHighlighted) }
+    </pre>
+  );
 };
 
 Code.propTypes = {
   /** The code within the component */
   children: React.PropTypes.string.isRequired,
+  /** Apply syntax highlighting to the code with automatic language detection */
+  isHighlighted: React.PropTypes.bool,
   /** Hook for automated JavaScript tests */
   testSection: React.PropTypes.string,
   /** How the code should be displayed */