+++ src/components/Code/tests/index.js
@@ -41,6 +41,42 @@ describe('components/Code', () => {
     expect(componentBlockNode.textContent).toBe(code);
   });
 
+  it('should add syntax highlighting to code when specified', () => {
+    let code = 'var foo;';
+
+    const componentInline = testHelpers.renderIntoDocument(
+      <Code type="inline" isHighlighted={ true }>{ code }</Code>
+    );
+
+    const componentBlock = testHelpers.renderIntoDocument(
+      <Code type="block" isHighlighted={ true }>{ code }</Code>
+    );
+
+    const componentInlineNode = testHelpers.getNodeFromComponent(componentInline);
+    const componentBlockNode = testHelpers.getNodeFromComponent(componentBlock);
+
+    expect(componentInlineNode.innerHTML).toContain('class="hljs-');
+    expect(componentBlockNode.innerHTML).toContain('class="hljs-');
+  });
+
+  it('should not add syntax highlighting by default', () => {
+    let code = 'var foo;';
+
+    const componentInline = testHelpers.renderIntoDocument(
+      <Code type="inline">{ code }</Code>
+    );
+
+    const componentBlock = testHelpers.renderIntoDocument(
+      <Code type="block">{ code }</Code>
+    );
+
+    const componentInlineNode = testHelpers.getNodeFromComponent(componentInline);
+    const componentBlockNode = testHelpers.getNodeFromComponent(componentBlock);
+
+    expect(componentInlineNode.innerHTML).not.toContain('class="hljs-');
+    expect(componentBlockNode.innerHTML).not.toContain('class="hljs-');
+  });
+
   it('should have a properly set test section', () => {
     const componentInline = testHelpers.renderIntoDocument(
       <Code testSection="foo-inline" type="inline">Hello!</Code>