+++ webpack.config.js
@@ -12,6 +12,10 @@ module.exports = {
         exclude: /(node_modules|bower_components)/,
         loader: 'babel',
       },
+      {
+        test: /\.css$/,
+        loader: 'style-loader!css-loader',
+      },
     ],
   },
 };